change: draft layout init
